agents-backend/pyproject.toml
[project]
name = "agents-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cerebras-cloud-sdk>=1.67.0",
    "fastapi>=0.135.1",
    "gitpython>=3.1.46",
    "google-generativeai>=0.5.0",
    "httpx>=0.28.1",
    "langchain>=0.3.0",
    "langchain-aws>=0.2.0",
    "langchain-cerebras>=0.8.2",
    "langchain-core>=1.2.17",
    "langchain-google-genai>=4.2.1",
    "langchain-groq>=1.1.2",
    "langchain-openai>=0.2.0",
    "langgraph>=1.0.10",
    "mcp>=1.26.0",
    "pandas>=3.0.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "python-dotenv>=1.2.2",
    "python-levenshtein>=0.27.3",
    "qdrant-client>=1.17.0",
    "scikit-learn>=1.8.0",
    "setuptools>=82.0.1",
    "tree-sitter>=0.25.2",
    "tree-sitter-java>=0.23.5",
    "unidiff>=0.7.5",
    "uvicorn>=0.41.0",
]
[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
